Fix mistake in 7edda82b where any key could delete the material

+++ pcbnew/board_stackup_manager/dialog_dielectric_list_manager.cpp
@@ -134,6 +134,6 @@ void DIALOG_DIELECTRIC_MATERIAL::onListKeyDown( wxListEvent& event )
 {
     int idx = event.GetIndex();
 
-    if( idx > 0 )
+    if( ( idx > 0 ) && ( event.GetKeyCode() == WXK_DELETE ) )
         m_lcMaterials->DeleteItem( idx );
 }
